3tier fixed

+++ src/3tier.py
@@ -12,7 +12,7 @@ import numpy as np
 if __name__ == '__main__':
     try:
         
-        data={"Cli":int(np.linspace(1,15,15)),"RTm":[],"rtCI":[],"Tm":[],"trCI":[],"ms":[]}
+        data={"Cli":np.linspace(1,15,15,dtype=int),"RTm":[],"rtCI":[],"Tm":[],"trCI":[],"ms":[]}
         
 
                # "ms3":{   "appFile":"../3tier/ms3.js",
